Refactor facet query getters
There are two separate facet query getters:
one for the full 'normal' facet query, and another
for the 'date-excluded' facet query that is used
to fetch the year histogram data.

The former of these used to build upon the latter,
but that setup didn't take into account that the
'date-excluded' query should also not include year
facets (it was only excluding the separate date ranges).

With that logic fixed, building one of these off the
other becomes a bit messy, so I've instead decoupled the
two queries and moved as much shared query-building
code out into separate functions for a reasonable
degree of DRYness (though there is still some minimal
repetition).